Seperating Components and Constants
HPWH system components moved to another file
Global contants and unit coversions moved to cfg.py
