+++ cfg.py
@@ -1,0 +1,8 @@
+
+
+# Declaring variables with a global scope
+
+rhoCp = 8.353535;
+W_TO_BTUHR = 3.412142;
+W_TO_BTUMIN = W_TO_BTUHR/60;
+W_TO_TONS = 0.000284345;
